Experiments with rewriting AutoValueProcessor and AutoOneOfProcessor to use JavaPoet. The principal purpose of doing this was to compare performance against the template-based approach that AutoValue has historically used. I expected that templates would be slower, but to my surprise they were about 10% faster. Possibly a more careful use of JavaPoet would allow it to equalize, but I would be surprised if it could be much faster than equality.
This code is not quite complete. Some functional tests fail, because I gave up before getting the copying of type annotations exactly right. (We need to respect exclusions in @CopyAnnotations.) Some functional tests had to be excluded because of limitations in JavaPoet concerning clashes with inherited nested types. GWT support doesn't work, but we don't really care about that.